+++ packages/insight-viewer/src/utils/common/getPolyProps.ts
@@ -5,6 +5,7 @@ interface GetPolyViewerInfoProps {
   annotation: PolygonAnnotation | LineAnnotation | FreeLineAnnotation
   selectedAnnotation: Annotation | null
   showOutline: boolean
+  pixelToCanvas: (point: Point) => Point
   annotationAttrs?: (annotation: Annotation, showOutline: boolean) => SVGProps<SVGPolygonElement>
 }
 
@@ -22,6 +23,7 @@ export function getPolyViewerInfo({
   showOutline,
   selectedAnnotation,
   annotationAttrs,
+  pixelToCanvas,
 }: GetPolyViewerInfoProps): getPolyViewerInfoReturnType {
   const { label, id, labelPosition: _labelPosition, points } = annotation
 
@@ -29,11 +31,11 @@ export function getPolyViewerInfo({
   const polygonLabel = label ?? id
 
   const polygonAttributes = typeof annotationAttrs === 'function' ? annotationAttrs(annotation, showOutline) : undefined
-  const labelPosition = _labelPosition || undefined
+  const labelPosition = _labelPosition ? pixelToCanvas(_labelPosition) : undefined
 
   const polygonPoints: string = points
     .map(point => {
-      const [x, y] = point
+      const [x, y] = pixelToCanvas(point)
       return `${x},${y}`
     })
     .join(' ')
@@ -42,7 +44,7 @@ export function getPolyViewerInfo({
     annotation.type === 'line' && annotation?.headPoints
       ? annotation.headPoints
           .map(point => {
-            const [x, y] = point
+            const [x, y] = pixelToCanvas(point)
             return `${x},${y}`
           })
           .join()