+++ packages/insight-viewer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.5.2](https://github.com/lunit-io/frontend-components/compare/@lunit/insight-viewer@5.5.1...@lunit/insight-viewer@5.5.2) (2022-05-02)
+
+
+### Bug Fixes
+
+* **viewer:** fix broken test ([3c0b1bb](https://github.com/lunit-io/frontend-components/commit/3c0b1bb12f879666d42b0fe00bfb7b6adceed02d))
+
+
+
+
+
 ## [5.5.1](https://github.com/lunit-io/frontend-components/compare/@lunit/insight-viewer@5.5.0...@lunit/insight-viewer@5.5.1) (2022-04-27)
 
 **Note:** Version bump only for package @lunit/insight-viewer